Introduced routes and templates

+++ MY_BANK_SYSTEM/MY_BANK_SYSTEM/account/models.py
@@ -7,7 +7,7 @@ from currency.models import Currency
 class Account(models.Model):
     accountNumber = models.CharField(max_length=200)
     currency = models.ForeignKey(Currency, on_delete=models.CASCADE)
-    balance = models.IntegerField()
+    balance = models.FloatField()
 
     class Meta:
         db_table = 'tbl_Account'
